requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
crashtest==0.4.1
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.1
distlib==0.3.6
distro==1.9.0
dm-tree==0.1.8
docutils==0.20.1
dulwich==0.21.5
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.98.0
fastjsonschema==2.19.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.12.0
gast==0.5.4
google==3.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
ipykernel==6.25.1
ipython==8.14.0
jaraco.classes==3.2.3
jedi==0.19.0
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
json5==0.9.14
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab_pygments==0.3.0
keras==2.15.0
keyring==23.13.1
kiwisolver==1.4.5
langchain==0.0.230
langchainplus-sdk==0.0.20
libclang==16.0.6
llama-cpp-python==0.1.70
lockfile==0.12.2
lxml==5.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.2.1
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
openai==1.50.2
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.6.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.16.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyproject_hooks==1.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline==2.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.1
PyYAML==6.0
pyzmq==24.0.1
rapidfuzz==2.15.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
sas7bdat==2.2.3
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.13.2
Send2Trash==1.8.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
statistics==1.0.3.5
sympy==1.12
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomlkit==0.11.8
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.5
traitlets==5.9.0
trove-classifiers==2023.5.24
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
